[feature-73/updating-ui-layer] Refactor in splash feature
