+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppThemeLight">
-        <activity android:name=".ui.activities.MainActivity">
+        <activity android:name=".ui.common.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
